+++ lib/rig/distributed_set.ex
@@ -176,6 +176,7 @@ defmodule RIG.DistributedSet do
     |> save(record)
     |> broadcast_update(pg2_group, last_record_id)
 
+    # increase Prometheus metrics (total, current) with an item in a set
     DistributedSetMetrics.add_item(name)
     {:reply, :ok, %{state | last_record_id: uuid}}
   end
@@ -211,6 +212,8 @@ defmodule RIG.DistributedSet do
           local_last_record_id
       end
 
+    # increase Prometheus metrics (total, current) with an item in a set
+    # TODO: maybe not needed
     DistributedSetMetrics.add_item(name)
     {:noreply, %{state | last_record_id: synced_record_id}}
   end
@@ -269,6 +272,7 @@ defmodule RIG.DistributedSet do
       Logger.debug(fn -> "Removed #{n_deleted} expired records" end)
     end
 
+    # decrease Prometheus metric (current) with an item not in a set
     DistributedSetMetrics.delete_item(name, n_deleted)
     n_deleted
   end